# 3x faster linear pathbuilder