--- a/bench/uPlot-10M.html
+++ b/bench/uPlot-10M.html
@@ -60,7 +60,7 @@
 				return data;
 			}
 
-			function makeChart(data) {
+			function makeChart(data, one = false) {
 				let series = [];
 
 				for (let i = 1; i < data.length; i++) {
@@ -80,7 +80,7 @@
 					},
 					cursor: {
 						points: {
-							one: true,
+							one,
 						},
 						focus: {
 							prox: 30
@@ -105,11 +105,12 @@
 
 		//	let data1 = prepData(10);
 		//	let data1 = prepData(5e6);
+		//	let data1 = prepData(1e7);
 			let data1 = prepData(2e6);
-			let u1 = makeChart(data1);
+			let u1 = makeChart(data1, false);
 
 			let data2 = prepData(40_000, 50, 2);
-			let u2 = makeChart(data2);
+			let u2 = makeChart(data2, true);
 
 			setTimeout(() => {
 				console.time('chart1 redraw');