--- a/src/paths/linear.js
+++ b/src/paths/linear.js
@@ -43,12 +43,6 @@ export function linear(opts) {
 			const _paths = {stroke: new Path2D(), fill: null, clip: null, band: null, gaps: null, flags: BAND_CLIP_FILL};
 			const stroke = _paths.stroke;
 
-			let minY = inf,
-				maxY = -inf,
-				inY, outY, outX, drawnAtX;
-
-			let accX = pixelForX(dataX[dir == 1 ? idx0 : idx1]);
-
 			// data edges
 			let lftIdx = nonNullIdx(dataY, idx0, idx1,  1 * dir);
 			let rgtIdx = nonNullIdx(dataY, idx0, idx1, -1 * dir);
@@ -57,52 +51,84 @@ export function linear(opts) {
 
 			let hasGap = false;
 
-			for (let i = dir == 1 ? idx0 : idx1; i >= idx0 && i <= idx1; i += dir) {
-				let x = pixelForX(dataX[i]);
-				let yVal = dataY[i];
+			// decimate when number of points >= 4x available pixels
+			const decimate = idx1 - idx0 >= xDim * 4;
 
-				if (x == accX) {
-					if (yVal != null) {
-						outY = pixelForY(yVal);
+			if (decimate) {
+				let xForPixel = pos => u.posToVal(pos, scaleX.key, true);
 
-						if (minY == inf) {
-							lineTo(stroke, x, outY);
-							inY = outY;
+				let minY = null,
+					maxY = null,
+					inY, outY, drawnAtX;
+
+				let accX = pixelForX(dataX[dir == 1 ? idx0 : idx1]);
+
+				let idx0px = pixelForX(dataX[idx0]);
+				let idx1px = pixelForX(dataX[idx1]);
+
+				// tracks limit of current x bucket to avoid having to get x pixel for every x value
+				let nextAccXVal = xForPixel(dir == 1 ? idx0px + 1 : idx1px - 1);
+
+				for (let i = dir == 1 ? idx0 : idx1; i >= idx0 && i <= idx1; i += dir) {
+					let xVal = dataX[i];
+					let reuseAccX = dir == 1 ? (xVal < nextAccXVal) : (xVal > nextAccXVal);
+					let x = reuseAccX ? accX :  pixelForX(xVal);
+
+					let yVal = dataY[i];
+
+					if (x == accX) {
+						if (yVal != null) {
+							outY = yVal;
+
+							if (minY == null) {
+								lineTo(stroke, x, pixelForY(outY));
+								inY = minY = maxY = outY;
+							} else {
+								if (outY < minY)
+									minY = outY;
+								else if (outY > maxY)
+									maxY = outY;
+							}
+						}
+						else {
+							if (yVal === null)
+								hasGap = true;
+						}
+					}
+					else {
+						if (minY != null)
+							drawAcc(stroke, accX, pixelForY(minY), pixelForY(maxY), pixelForY(inY), pixelForY(outY));
+
+						if (yVal != null) {
+							outY = yVal;
+							lineTo(stroke, x, pixelForY(outY));
+							minY = maxY = inY = outY;
+						}
+						else {
+							minY = maxY = null;
+
+							if (yVal === null)
+								hasGap = true;
 						}
 
-						minY = min(outY, minY);
-						maxY = max(outY, maxY);
-					}
-					else {
-						if (yVal === null)
-							hasGap = true;
+						accX = x;
+						nextAccXVal = xForPixel(accX + dir);
 					}
 				}
-				else {
-					if (minY != inf) {
-						drawAcc(stroke, accX, minY, maxY, inY, outY);
-						outX = drawnAtX = accX;
-					}
 
-					if (yVal != null) {
-						outY = pixelForY(yVal);
-						lineTo(stroke, x, outY);
-						minY = maxY = inY = outY;
-					}
-					else {
-						minY = inf;
-						maxY = -inf;
+				if (minY != null && minY != maxY && drawnAtX != accX)
+					drawAcc(stroke, accX, pixelForY(minY), pixelForY(maxY), pixelForY(inY), pixelForY(outY));
+			}
+			else {
+				for (let i = dir == 1 ? idx0 : idx1; i >= idx0 && i <= idx1; i += dir) {
+					let yVal = dataY[i];
 
-						if (yVal === null)
-							hasGap = true;
-					}
-
-					accX = x;
+					if (yVal === null)
+						hasGap = true;
+					else if (yVal != null)
+						lineTo(stroke, pixelForX(dataX[i]), pixelForY(yVal));
 				}
 			}
-
-			if (minY != inf && minY != maxY && drawnAtX != accX)
-				drawAcc(stroke, accX, minY, maxY, inY, outY);
 
 			let [ bandFillDir, bandClipDir ] = bandFillClipDirs(u, seriesIdx);
 
@@ -116,7 +142,7 @@ export function linear(opts) {
 				lineTo(fill, lftX, fillToY);
 			}
 
-			if (!series.spanGaps) {
+			if (!series.spanGaps) { // skip in mode: 2?
 			//	console.time('gaps');
 				let gaps = [];